move plotting_utilities from feec multipatch to fem
